% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelrwa.R
\name{include_interactions}
\alias{include_interactions}
\title{Include interactions in a model}
\usage{
include_interactions(
  formula = NULL,
  response.name = NULL,
  data = NULL,
  pos.control = NULL,
  pos.fixed = NULL,
  pos.variables = NULL,
  include.interactions = FALSE,
  family,
  alpha = 0.01,
  method = c("aic", "p")
)
}
\arguments{
\item{formula}{an object of class formula with main effects.}

\item{response.name}{output name of the data.}

\item{data}{a data.frame.}

\item{pos.control, pos.fixed, pos.variables}{position of the control, fixed and variables inputs in the data.frame.}

\item{include.interactions}{a boolean indicating if create or no a pairwise set of interactions.}

\item{family}{type of regression (see \code{\link[stats]{family}})}

\item{alpha}{level of significance use if \code{method = "p"}.}

\item{method}{method used to keep interactions if \code{include.interactions=TRUE}. See \code{\link[rms]{fastbw}}.}
}
\value{
a list with the control, fixed, variables and interactions used in the model. Also it returns the the fitted model.
}
\description{
Include interactions in a model
}
\keyword{internal}
