% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residualrwa.R
\name{plot.residualrwa}
\alias{plot.residualrwa}
\title{Plot for residualrwa objects}
\usage{
\method{plot}{residualrwa}(x, boot_ci, font_size = 12, ...)
}
\arguments{
\item{x}{An object of class \code{"residualrwa"}.}

\item{boot_ci}{A boolean. Its value determines if the function draws or not
the boostrap confidence. If misssing, then the function uses the value in
\code{x$boot_ci}.}

\item{font_size}{A numeric value. Overall font size. It uses the same
mechanism as \code{\link[cowplot]{theme_cowplot}}.}

\item{...}{Currently ignored.}
}
\value{
Returns a ggplot object
}
\description{
Plot for residualrwa objects
}
\examples{

n <- 100
X1 <- rnorm(n)
X2 <- rnorm(n)
X3 <- rnorm(n)
Y <- X2^3 + 10 * X1 * X2
data <- as.data.frame(cbind(Y, X1, X2, X3))

ex <- residualrwa(
  response = "Y",
  control = NULL,
  fixed = NULL,
  free = c("X1", "X2", "X3"),
  data = data,
  include_interactions = TRUE,
  boot_ci = TRUE,
  n_boot = 5
)

plot(ex)

}
