% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_residualrwa.R
\name{print.residualrwa}
\alias{print.residualrwa}
\title{Print residualrwa objects}
\usage{
\method{print}{residualrwa}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"residualrwa"}.}

\item{...}{Currently ignored.}
}
\value{
Print summarized information about the \code{residualrwa} object.
}
\description{
Print residualrwa objects
}
\examples{

n <- 100
X1 <- rnorm(n)
X2 <- rnorm(n)
X3 <- rnorm(n)
Y <- X2^3 + 10 * X1 * X2
data <- as.data.frame(cbind(Y, X1, X2, X3))

ex <- residualrwa(
  response = "Y",
  control = NULL,
  fixed = NULL,
  free = c("X1", "X2", "X3"),
  data = data,
  include_interactions = TRUE,
  boot_ci = TRUE,
  n_boot = 5
)

print(ex)

}
