% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualrwa.R
\name{residualrwa}
\alias{residualrwa}
\title{Nonlinear relative weight analysis with residualization}
\usage{
residualrwa(
  response,
  control = NULL,
  fixed = NULL,
  free,
  data,
  family = stats::gaussian(),
  include_interactions = FALSE,
  name_control = "Control",
  name_fixed = "Fixed",
  name_free = "Free",
  name_interactions = "Interactions",
  boot_ci = FALSE,
  n_boot = 100,
  mc_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{response, free, fixed, control}{Character or character vectors with names
for each parameter. Used in \code{residualrwa} to determine how to treat
each variable. \code{response} is the model output. \code{free} are
variables that can be included or removed (even in interactions) in all the
models without restrictions. \code{fixed} are persistent variables across
all the models, however they can be included/removed on interactions.
\code{control} they only act as main effects and never are used for
interactions.}

\item{data}{A \code{data.frame}.}

\item{family}{A specification for the model link function. See
\code{\link[stats]{family}} for accepted values.}

\item{include_interactions}{A boolean with default \code{FALSE}. Determine if
the model should calculate all the pairwise interactions between variables.
It uses the names in the parameters \code{free} and \code{fixed}.}

\item{name_free, name_fixed, name_control, name_interactions}{A string for type
of variable with defaults "Free", "Fixed", "Control" and "Interaction".
Names used to label the summary tables}

\item{boot_ci}{A boolean with default \code{FALSE}. Determine if a bootstrap
procedure should be used to estimate confidence intervals for the weights.
Defaults to \code{FALSE}.}

\item{n_boot}{A numeric with default 100. Number of bootstrap samples used to
estimate the confidence intervals.}

\item{mc_cores}{A numeric with default 1. Number of cores used when
performing the bootstrap samples.}

\item{verbose}{A boolean with default \code{FALSE}. If \code{TRUE},
\code{residualrwa} shows message about the stepwise process.}
}
\value{
A residualrwa object with this structure:

\describe{

\item{summary}{a \code{data.frame} with the consolidated sum of relative
weights according to each type of component; free, fixed, control and
interactions.}

\item{data_frame}{a \code{data.frame} with the individual relative weights
for each variable. The \code{data.frame} has columns: \code{variable},
\code{weight} and \code{type}. If \code{boot_ci = TRUE}, two optional columns
\code{ci_low} and \code{ci_up} indicating the lower and upper confidence
intervals for each variable respectively.}

\item{model}{a \code{\link[rms]{Glm}} object with the final model used to
estimate the relative weights.}

\item{data}{original \code{data.frame}.}

\item{variables}{a list with character vectors \code{free}, \code{fixed},
\code{control} and \code{interactions} with the final variables used in the
model.}

\item{boot_ci}{a boolean indicating if the bootstrap procedure was used.}

}
}
\description{
Method to detecting influential variables in nonlinear model
  with residualized relative weight analysis.
}
\examples{

n <- 100
X1 <- rnorm(n)
X2 <- rnorm(n)
X3 <- rnorm(n)
Y <- X2^3 + 10 * X1 * X2
data <- as.data.frame(cbind(Y, X1, X2, X3))

ex <- residualrwa(
  response = "Y",
  control = NULL,
  fixed = NULL,
  free = c("X1", "X2", "X3"),
  data = data,
  include_interactions = TRUE,
  boot_ci = TRUE,
  n_boot = 5
)

}
