% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{RWA}
\alias{RWA}
\title{Relative Weight Analysis for logistic regression}
\usage{
RWA(X, Y, data, family)
}
\arguments{
\item{X}{matrix or data.frame with inputs}

\item{Y}{numeric vector or data.frame with output.}

\item{data}{data.frame}
}
\value{
a list with the R^2, adjusted R^2, and relative weights for the X with respect to Y.
}
\description{
Relative Weight Analysis for logistic regression
}
\keyword{internal}
