% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualrwa.R
\name{residualrwa}
\alias{residualrwa}
\title{#' Detect influential variables in model with Relative Weight Analysis}
\usage{
residualrwa(
  response.name,
  control = NULL,
  fixed = NULL,
  free,
  data,
  family = stats::gaussian(),
  include.interactions = FALSE,
  alpha = 0.1,
  method = c("aic", "p"),
  name.control = "Control",
  name.fixed = "Fixed",
  name.free = "Free",
  name.interactions = "Interactions",
  verbose = FALSE
)
}
\arguments{
\item{response.name}{name for independent variable}

\item{control}{character vector with names for control variables}

\item{fixed}{character vector with names for fixed variables}

\item{free}{character vector with names for free variables}

\item{data}{object data.frame with data}

\item{family}{type of regression (see \code{\link[stats]{family}})}

\item{include.interactions}{A boolean indicating if the model should calculate all the pairwise interactions between variables. It uses the names in the parameters \code{fixed} and \code{variables}.}

\item{alpha}{level of significance use if \code{method = "p"}.}

\item{method}{method used to keep interactions if \code{include.interactions=TRUE}. See \code{\link[rms]{fastbw}}.}

\item{name.control, name.fixed, name.variables, name.interactions}{Names used to label the summary tables}

\item{verbose}{if \code{TRUE}, \code{residualrwa} shows all the stepwise process. Defaults to \code{FALSE}.}
}
\value{
A residualrwa object
}
\description{
#' Detect influential variables in model with Relative Weight Analysis
}
\examples{

X1 <- rnorm(1000)
X2 <- rnorm(1000)
X3 <- rnorm(1000)

Y <- plogis(X2^3+10*X1*X2)>0.5

data <- as.data.frame(cbind(Y,X1,X2,X3))

ex1 <- residualrwa(
  response.name = "Y" ,
  control = NULL,
  fixed = NULL,
  variables = c("X1","X2", "X3"),
  data = data,
  family = binomial,
  include.interactions = TRUE
)




}
